waiting list implemented. some issues fixed, but not all

+++ src/Adopt.js
@@ -1,4 +1,5 @@
 import React, { Component } from "react";
+import uuid from 'react-uuid'
 import Context from "./Context";
 import Cats from "./Cats/Cats";
 import Dogs from "./Dogs/Dogs";
@@ -8,7 +9,6 @@ import dogService from "./services/dogServices";
 
 export default class Home extends Component {
   state = {
-    //cats: [],
     currentCat: null,
     currentDog: null,
     currentAdoptedPet: null,
@@ -20,7 +20,9 @@ export default class Home extends Component {
 
   componentDidMount() {
     this.context.loadUsers();
-
+    if (!this.context.users[0]) {
+      this.props.history.push("/");
+    }
     this.getNextCat();
     this.getNextDog();
 
@@ -39,13 +41,29 @@ export default class Home extends Component {
       }
     }, this.context.stopTime);
   }
+
   componentWillUnmount() {}
+
+  getNextCat() {
+    catService.getNextCat().then((currentCat) => {
+      this.setState({
+        currentCat: currentCat,
+      });
+    });
+  }
+
+  getNextDog() {
+    dogService.getNextDog().then((currentDog) => {
+
+      this.setState({
+        currentDog: currentDog,
+      });
+    });
+  }
 
   adoptCat = () => {
     this.setState(
-      {
-        currentAdoptedPet: this.state.currentCat,
-      },
+      {currentAdoptedPet: this.state.currentCat},
       function () {
         catService.deleteCurrent().then(() => {
           setTimeout(() => {
@@ -57,9 +75,7 @@ export default class Home extends Component {
   };
   adoptDog = () => {
     this.setState(
-      {
-        currentAdoptedPet: this.state.currentDog,
-      },
+      {currentAdoptedPet: this.state.currentDog},
       function () {
         dogService.deleteCurrent().then(() => {
           setTimeout(() => {
@@ -79,8 +95,10 @@ export default class Home extends Component {
       function () {
         catService.deleteCurrent().then(() => {
           setTimeout(() => {
-            console.log("redirect to the home page");
-            this.props.history.push("/");
+            //this.props.history.push("/");
+            userService.deleteCurrent();
+            this.context.removeUser()
+            this.context.setUser()
           }, 2000);
         });
       }
@@ -95,8 +113,10 @@ export default class Home extends Component {
       function () {
         dogService.deleteCurrent().then(() => {
           setTimeout(() => {
-            console.log("redirect to the home page");
-            this.props.history.push("/");
+            //this.props.history.push("/");
+            userService.deleteCurrent();
+            this.context.removeUser();
+            this.context.setUser();
           }, 2000);
         });
       }
@@ -111,8 +131,10 @@ export default class Home extends Component {
         dogService.deleteCurrent().then(() => {
           catService.deleteCurrent().then(() => {
             setTimeout(() => {
-              console.log("redirect to the home page");
-              this.props.history.push("/");
+              //this.props.history.push("/");
+              userService.deleteCurrent();
+              this.context.removeUser()
+              this.context.setUser()
             }, 2000);
           });
         });
@@ -124,7 +146,7 @@ export default class Home extends Component {
     dogService.getNextDog().then((currentDog) => {
       this.context.loadUsers();
       this.setState({
-        currentDog: currentDog,
+        currentDog,
         currentAdoptedPet: null,
       });
     });
@@ -134,46 +156,37 @@ export default class Home extends Component {
     catService.getNextCat().then((currentCat) => {
       this.context.loadUsers();
       this.setState({
-        currentCat: currentCat,
+        currentCat,
         currentAdoptedPet: null,
       });
     });
   };
 
   getRandomInt = (max) => {
-    return Math.floor(Math.random() * Math.floor(max)); //60
+    return Math.floor(Math.random() * max); //60
   };
 
-  //need to clear interval from set interval with componenet did unmount
+
+  //need to clear interval from set interval with component did unmount
   shuffleUser = () => {
-    if (this.context.users.length) {
-      //it means that you are the only user left in the queue
-      //so you are in front of the line so you start adding users
+    if (this.context.users.length && this.context.user !== this.context.users[0]) {
       userService.deleteCurrent().then(() => {
-        if (this.getRandomInt(100) < 50) {
+        if (this.getRandomInt(10) <= 5) {
+          if(this.state.currentCat){
           this.adoptCat();
+          } else {
+            this.adoptDog();
+          }
         } else {
-          this.adoptDog();
+          if(this.state.currentDog){
+            this.adoptDog();
+            } else {
+              this.adoptCat();
+            }
         }
       });
     }
   };
-
-  getNextCat() {
-    catService.getNextCat().then((currentCat) => {
-      this.setState({
-        currentCat: currentCat,
-      });
-    });
-  }
-
-  getNextDog() {
-    dogService.getNextDog().then((currentDog) => {
-      this.setState({
-        currentDog: currentDog,
-      });
-    });
-  }
 
   getCurrentUserStatus = () => {
     if (this.state.bothPetAdoptionStatus) {
@@ -213,14 +226,6 @@ export default class Home extends Component {
       </button>
     );
   };
-  //   addToQueue = () => {
-  //     while(this.context.users.length < 5){
-  //         setInterval(() => {
-  //             userService.postUser('random')
-  //             this.context.loadUsers()
-  //         }, 5000);
-  //     }
-  //   };
 
   render() {
     let value = {
@@ -233,7 +238,8 @@ export default class Home extends Component {
     };
     if (this.context.isLoading) {
       return <p>Loading</p>;
-    } else {
+    }
+    else {
       return (
         <Context.Provider value={value}>
           <div>
@@ -244,6 +250,8 @@ export default class Home extends Component {
             {!this.state.bothPetAdoptionStatus && <Cats className="pet" />}
             {!this.state.bothPetAdoptionStatus && <Dogs className="pet" />}
             {this.getAdoptBothButton()}
+            <h3>Currently Waiting:</h3>
+            {this.context.users.map((user) => {return <p key={uuid()}>{user}</p>})}
           </div>
         </Context.Provider>
       );
